Make `checkDomain` fail when a non-Zulip URL is entered.
Before, any URL without a status code response of 404 and without an
`error` in the `request` callback would fail. However, `error` doesn't
actually populate for error responses; it is populated when there is an
error with the request. Therefore, many websites that responded with
non-404 (but still error) status codes would be accepted. Instead, check
for all error status codes as well as the `request` error.