reconnect: Check wheather internet is working before reloading.
When online event is triggered check weather the internet is atually working or not.
Commonly on windows it turns out that internet takes couple of seconds to boot up after
connecting to internet or in some cases this might be they have to sign in to internet service
portal in order to access internet.