Altered path mentioned in defaultIconUrl
The icon previously was linked to the the path
'../renderer/img/icon.png' which made it only possible for files having
their working dirctory in the same directory as that of 'renderer' to
access the file via the fs. Ideally one would like the path to be set
such that the fallback image can be accessed effectively from any part
of the repository. Using '__dirname' we get the working directory from
'utils', slice it to backtrack till render and add the icon path from
there onwards. This gives a global path back to the image, one that can
be invoked from anywhere in the repository, if not the system to get the
fallback image.