Altered path mentioned in defaultIconUrl
The icon previously was linked to the the path
'../renderer/img/icon.png' which made it only possible for files having
their working dirctory in the same directory as that of 'renderer' to
access the file via the fs. Ideally one would like the path to be set
such that the fallback image can be accessed effectively from any part
of the repository. Using '__dirname' and 'path.resolve()', we get a
universal path to the default icon so the fall back goes as intended.