sentry-util: Hard-code the Sentry DSN.
Commit 088ddf9c62d1a181b88da25c45c1d996250bfaba (#755) does not work,
because neither the .env file nor the environment variables it
provides are available to normal users at runtime.  This silently
broke Sentry data collection.  When we upgraded @sentry/electron in
commit 107e5229141879b69225153dc9e6d15f54c8c7bf, the silent failure
became an error that prevented the app from starting.

The Sentry DSN is not a secret, so we should just commit it to the
repository.

Signed-off-by: Anders Kaseorg <anders@zulipchat.com>