report-issue: Remove unnessacary instructions from the placeholder.
It looks like the removed instructions were directly copied from GitHub
issue template and did not fit here. Specifically, the first point about
including the platform is automatically included in the email we prepare for
the user and we also include browser and electron version. Then the second
bullet point re-iterates the instructions present. Lastly, the third point
asks for screenshot however user cannot add screenshot in the UI but they can
in the email.

We also remove an comment disabling a eslint rule. The rule was disable for no
reason because we should have used template literal instead of the undocumented
multi-str format which is not cleaner.