UX: Font size for spellchecker explanation decreased.
The font size of explanation text was as same as title so i have changed its size to x-small for better UX.

UX: Font size for spellchecker explanation decreased.

The font size of explanation text was as same as title so i have changed its size to x-small for better UX.

UX: Font size for spellchecker explanation decreased.

Fixes #978 .The font size of explanation text was as same as title so i have changed its size to x-small for better UX.

UX: Font size for spellchecker explanation decreased.

Fixes #978 .The font size of explanation text was as same as title so i have changed its size to 10px for better UX.