darwin-notifications: Migrate darwin-notifications to electron's native Notifications.
Electron's native Notifications now supports inline replies on macOS and thus we don't need any 3rd party package.
This commit removes the usage of node-mac-notifier and brings native Notification in picture.

For comprehensibility, ipc-helpers will contain the functions which make use of Main process APIs with already functioning code in Renderer process.
This is a potential solution to people not getting notifications on macOS.