tab: Use activeTab instead of activeTabIndex in ServerManagerView.
Tracking these indices separately than the tab object means that we will
have low confidence trying to control the order by changing the indices.
Using the Tab object instead as a reference is the right way to go about
this.

This commit only changes activeTabIndex to activeTab for
ServerManagerView. We will change activeTabIndex to activeTab at other
applicable places in further commits to keep commits more readable.
getActiveTabIndex is a temporary function to help us solve this, it
should be removed once our work is done.