tab: Use activeTab instead of activeTabIndex in menu.ts.
Tracking these indices separately than the tab object means that we will
have low confidence trying to control the order by changing the indices.
Using the Tab object instead as a reference is the right way to go about
this.

This commit marks the completion of activeTabIndex removal.
getNextServer and getPreviousServer still pass an index to
switch-server-tab but that will be handled on it's own in `index`
removal in favour of the Tab object.