domain: Add id field to server objects.
This id will be passed around in events to uniquely identify a server.
This id is not user generated, it is generated in app by domain-util.ts
and stored in our JSON DB. This commit just adds the logic for
generating and storing the id in the JSON DB. This commit does not make
use of this newly added id, that will be done in further commits to keep
the commits readable.

In further commits, we will also start setting tabId in the rendered
code to this id. We generate a random hex string for this id instead of
having a numeric incremental ID. The reason for this is to avoid
overhead of maintaining an ever increasingly numeric id while trying to
make sure that ID collisions don't happen. Since we don't store our data
in a real database, we can't have incremental primary keys by default
and it makes little sense to implement that on our own, when this field
is just being passed around to communicate b/w code and is not shown to
humans.