ipc: Use id for `switch-server-tab` and set tabId to server id.
We want to start using server id to communicate the identity of server
in events. This commit does that for `switch-server-tab`. It also
requires us to set tabId to server id instead of an incremental number.
We try to make changes in this commit minimal and pass index to the
functions that switch-server-tab calls. We will replace that index with
tabId in further commits incrementally.