tab: Use activeTabId instead of activeTabIndex.
We also use lastActiveTabId instead of lastActiveTabIndex now.
`save-last-tab` uses id instead of index to commmunicate the identity of
a server. getNextServer and getPreviousServer also start using id with
this commit, the index mentioned in those functions is actually the
order in which the servers should be displayed and we will fix that
problem in upcoming commits.