tab: Rename index to order.
After introducing id field to the server object, index was just an
indicator of the ordering of tabs in the tab list. This commit changes
the name to accurately represent the field's function. This commit
should make it easier to change the order of tabs in the future.

Since order is not the same as index anymore, we add an additional
function called getTabByOrder to replace the leftover instances of us
fetching a tab by it's index when it meant to fetch a tab by it's id.

After this commit, we should not have any instances of accessing tabs
directly by their index in the server list or the tab list.