download: Show in-app notification on downlaoding a file.
Fixes #1371.
We do not accept a filePath from the web app when recieving the event to
show the file in a folder, since that could introduce securithy
vulnerabilies.

We instead keep the list of latest 50 downloaded files in memory and
give each filePath a unique downloadId that can be passed around between
the webapp and the desktop app.